requirements.txt
cython
attr==0.3.2
contextlib2==21.6.0
cryptography==44.0.3
docutils==0.21.2
docx==0.2.4
google_api_python_client==2.167.0
HTMLParser==0.0.2
imap_tools==1.10.0
importlib_metadata==8.0.0
ipython==8.18.1
ipywidgets==8.1.7
Jinja2==3.1.6
keyring==25.6.0
langchain==0.3.25
langchain_google_genai==2.1.4
lockfile==0.12.2
mock==5.2.0
Pillow==11.2.1
protobuf==6.31.0
pydrive==1.3.1
pyOpenSSL==25.0.0
PyPDF2==3.0.1
python-dotenv==1.1.0
python_docx==1.1.2
railroad==0.5.0
Sphinx==8.2.3
thread==2.0.5
tornado==6.4.2
truststore==0.10.1
urllib3_secure_extra==0.1.0

